// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  CONTENT_MANAGER: 'CONTENT_MANAGER',
  VIEWER: 'VIEWER',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BLOCKED: 'BLOCKED',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const PaymentStatus = {
  PAID: 'PAID',
  UNPAID: 'UNPAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED',
  PENDING: 'PENDING'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const AccessType = {
  FREE: 'FREE',
  SUBSCRIPTION: 'SUBSCRIPTION',
  TICKET: 'TICKET',
  BOTH: 'BOTH'
} as const

export type AccessType = (typeof AccessType)[keyof typeof AccessType]


export const SubscriptionPlan = {
  FREE: 'FREE',
  PREMIUM_MONTHLY: 'PREMIUM_MONTHLY',
  PREMIUM_YEARLY: 'PREMIUM_YEARLY'
} as const

export type SubscriptionPlan = (typeof SubscriptionPlan)[keyof typeof SubscriptionPlan]


export const ReviewStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type ReviewStatus = (typeof ReviewStatus)[keyof typeof ReviewStatus]


export const MediaType = {
  MOVIE: 'MOVIE',
  TV_SERIES: 'TV_SERIES'
} as const

export type MediaType = (typeof MediaType)[keyof typeof MediaType]


export const PaymentPurpose = {
  TICKET_PURCHASE: 'TICKET_PURCHASE',
  SUBSCRIPTION_PURCHASE: 'SUBSCRIPTION_PURCHASE',
  SUBSCRIPTION_RENEWAL: 'SUBSCRIPTION_RENEWAL',
  REFUND: 'REFUND'
} as const

export type PaymentPurpose = (typeof PaymentPurpose)[keyof typeof PaymentPurpose]